My branch

+++ Backend_ProjetTicket_OCP/src/main/java/com/example/Backend_ProjetTicket_OCP/Repo/EquipeRepository.java
@@ -3,5 +3,5 @@ package com.example.Backend_ProjetTicket_OCP.Repo;
 import com.example.Backend_ProjetTicket_OCP.Entite.Equipe;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EquipeRepository extends JpaRepository<Equipe, Long> {
+public interface EquipeRepository extends JpaRepository<Equipe, Integer> {
 }
